requirements.txt
# Azure ML AutoML Model Builder - Requirements
# Updated: January 2026

# Core Azure ML packages
azureml-core==1.56.0
azureml-train-automl-client==1.56.0
azureml-train-automl-runtime==1.56.0
azureml-automl-core==1.56.0
azureml-automl-runtime==1.56.0
azureml-dataset-runtime==1.56.0
azureml-defaults==1.56.0
azureml-interpret==1.56.0
azureml-mlflow==1.56.0
azureml-pipeline-core==1.56.0
azureml-telemetry==1.56.0
azureml-train-core==1.56.0

# Data processing and ML
numpy>=1.23.5,<2.0.0
pandas>=2.0.3,<3.0.0
scikit-learn>=1.3.2,<2.0.0
xgboost>=2.0.3
lightgbm>=4.1.0

# MLflow
mlflow>=2.9.2
mlflow-skinny>=2.9.2

# Model interpretation
interpret-community>=0.31.0
shap>=0.44.0

# Azure services
azure-identity>=1.15.0
azure-storage-blob>=12.19.0
azure-mgmt-resource>=23.0.1
azure-mgmt-storage>=21.1.0

# Inference and deployment
inference-schema>=1.5.1
joblib>=1.3.2

# Time series (if needed)
prophet>=1.1.5
statsmodels>=0.14.1

# Utilities
python-dateutil>=2.8.2
pytz>=2023.3
requests>=2.31.0
urllib3>=2.1.0

# Development tools (optional)
black>=23.12.1
flake8>=7.0.0
mypy>=1.8.0
pytest>=7.4.4
pytest-cov>=4.1.0
ipykernel>=6.28.0
jupyter>=1.0.0

# Type stubs
types-python-dateutil>=2.8.19
types-requests>=2.31.0
